feat: refactor fetchUser methode
